Moved table_version.py
