+++ domain-ee/ee-max-cfi-app/src/python_src/util/data/table_version.py
